requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pymongo==4.6.0
python-dotenv==1.0.0
pandas==2.1.1
openpyxl==3.1.2
httpx==0.25.0
langchain==0.2.16
langchain-community==0.2.17
langchain-huggingface==0.0.3
sentence-transformers==2.7.0
chromadb==0.4.18
python-multipart==0.0.6
openai==1.46.1
httpx==0.27.0
sentence-transformers==3.2.0
transformers==4.44.2
torch==2.4.1
langchain-chroma==0.1.4